Refactor the hook to load details
We can't load them until the map loads. But before if the map loaded
before the details they would miss their chance and not load until
backing out and coming back to the object.

When the map loads initially set a `mapLoaded` variable, check for it
when the crashpad JSON loads, and maintain the `viewChanged` hook.