Flip the goldweights upside down
This is a bit sketchy, but it works. I'm using a css transform to flip
the entiry page 180º and then another 180º on the image of the table so
that it's oriented correctly. Then when the iPad is upside down, on the
west side of the installed table, it looks like the image of the table
is oriented correctly to the table.