refactor cluster/show all toggling logic
mutating scope.all sucked and was hard to keep track of. So at the outset,
scope up two sets: 'this cluster' and 'all objects', then toggle between
them.

The one downside here is that angular flashes when it reloads worse than
it did before because we're toggling to a different set. even though the
order of the objects in the first cluster remains the same, they've been
copied so they're different.

Maybe instead of 'cluster' and 'all' sets I could use 'cluster' and
'other', show/hiding the others instead of swapping between cluster and
all? TODO