black-translucent is the least buggy in iOS8
default gives me a white bar, often with white text. black has a
translucent status bar with white text and adds a 20px black bar to the
bottom of the screen unless it's launched in portrait mode and then
switched!?

black-translucent gives a translucent status bar with black text on
first launch and then translucent with white text the rest of the
launches.